fix: small refactoring
